requirements.txt
Flask==3.0.0
pymongo==4.7.2
requests==2.31.0
opentelemetry-distro[otlp]==0.45b0